Stage 2 of 5 for JetBrains Academy - Java - Battleship project. Description
The goal of this game is to sink all the ships of your opponent. Our fleet is not ready for a big battle yet, so let's practice shooting on our field. Place all your units on the battlefield and take a shot!

In this step, you need to develop a system of shooting with accompanying messages about hits and misses.

Objectives

Take a shot at a prepared game field. You need to indicate the coordinates of the target, and the program should then display a message about a hit or a miss. If the shell misses the target and falls in the water, this cell should be marked with an M, and a successful strike is marked by an X. After this shot, the game should be stopped.

Main.java changes:
• main() -  added a call to the new method startGame().
• startGame() - allows the user to enter one coordinate to launch a shot. It uses printOceanField() to print the field as required, getString(), testPosition(), and notRange() to verify coordinate, and then when confirmed informing the user that they either missed or hit a shipped depending on the coordinate entered; if not confirmed looping until it is.
• getPositions() - was split up so that startGame() can verify one coordinate. The testing of one coordinate was moved to testPosition().
• testPosition() - receives one string to verify it can be converted to two int numbers, and then returning them in an int array.