Stage 3 of 5 for JetBrains Academy - Java - Battleship project. Description
It seems a little odd to shoot your own ships. Let's imagine they are not ours! You can ask a friend to place the ships on the game field (or do it yourself, but it will be less exciting), and then the program will hide the ships under the fog of war. You just have to take a blind shot.
Objectives

In this stage, you need to implement the "fog of war" feature in your game. First, place all the ships on the game field, and then hide them with the symbol ~. Take a shot like in the previous stage, and after your attempt, the program should print a message along with two versions of the field: one covered with the fog of war and the other one uncovered.

Main.java changes:
• main(), placeShips() - adjusted them to account for changes made to printOceanField()
• printOceanField() - added a fog boolean to its calling method to adjust its printing if there is fog of war. Adjusted the inner for loop to check if fog is true and if so it hides the ships from view.
• startGame() - adjusted for changes made to printOceanField(), and the additional printing of the ocean field.